`utils::primes` output does not have to be case to `i64`.

+++ src/solutions/reciprocal_cycles.rs
@@ -7,7 +7,7 @@ pub fn solve() {
     // decimal recurrence. Exclude 3 because it is definitely not the answer.
     let result = utils::primes(1000)
         .skip_while(|prime| *prime <= 5)
-        .map(|prime| (utils::recurrence_length(prime as i64), prime))
+        .map(|prime| (utils::recurrence_length(prime), prime))
         .max()
         .unwrap()
         .1;
